# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenrouterConversation {
  id: number;
  title: string;
  model: string;
  category?: string;
  createdAt: string;
}

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenrouterConversationInput {
  title: string;
  model: string;
  category?: string;
}

export interface OpenrouterMessageInput {
  content: string;
}

export interface OpenrouterConversationWithMessages {
  id: number;
  title: string;
  model: string;
  category?: string;
  createdAt: string;
  messages: OpenrouterMessage[];
}

export interface OpenrouterError {
  error: string;
}