# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: openrouter
    description: OpenRouter AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /openrouter/conversations:
    get:
      operationId: listOpenrouterConversations
      tags: [openrouter]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterConversation"
    post:
      operationId: createOpenrouterConversation
      tags: [openrouter]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenrouterConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversation"
  /openrouter/conversations/{id}:
    get:
      operationId: getOpenrouterConversation
      tags: [openrouter]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
    delete:
      operationId: deleteOpenrouterConversation
      tags: [openrouter]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenrouterError"
  /openrouter/conversations/{id}/messages:
    get:
      operationId: listOpenrouterMessages
      tags: [openrouter]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenrouterMessage"
    post:
      operationId: sendOpenrouterMessage
      tags: [openrouter]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenrouterMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OpenrouterConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        model:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - model
        - createdAt
    OpenrouterMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    OpenrouterConversationInput:
      type: object
      properties:
        title:
          type: string
        model:
          type: string
        category:
          type: string
      required:
        - title
        - model
    OpenrouterMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenrouterConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        model:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenrouterMessage"
      required:
        - id
        - title
        - model
        - createdAt
        - messages
    OpenrouterError:
      type: object
      properties:
        error:
          type: string
      required:
        - error